fix: historyWindowを改修
