+++ src/ecosystem/history/index.module.scss
@@ -1,0 +1,53 @@
+.inner {
+  display: flex;
+  flex-direction: row;
+  width: 100%;
+}
+
+.image_block {
+  width: 55%;
+  margin-right: 5%;
+}
+
+.thumbnail_block {
+  width: 40%;
+  display: flex;
+  flex-direction: column;
+}
+
+.title {
+  text-align: center;
+  font-size: min(3.5vw, 45px);
+  margin-bottom: 15px;
+
+  @media screen and (max-width: 1024px) {
+    font-size: 8vw;
+    letter-spacing: 5px;
+    margin-bottom: 70px;
+  }
+}
+
+.content {
+  text-align: center;
+  font-size: min(1.8vw, 27px);
+
+  @media screen and (max-width: 1024px) {
+    font-size: 4vw;
+  }
+}
+
+.button_wrapper {
+  position: relative;
+  display: flex;
+  justify-content: center;
+  margin: 8% 0;
+}
+
+.thumbnail_line {
+  display: flex;
+  flex-direction: row;
+}
+
+.scroll_wrapper {
+  text-align: center;
+}
