+++ src/pages/index.module.scss
@@ -1,0 +1,4 @@
+.inner {
+  width: 80%;
+  margin: auto;
+}
